src/requirements.txt
amqp==5.0.2
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.3
Django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
gunicorn==20.0.4
kombu==5.0.2
mccabe==0.6.1
Pillow==8.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
